refactor: refactor home, search, login and registration pages
